% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{cgsolve}
\alias{cgsolve}
\title{Solve for x in Ax = b using conjugate gradient method.}
\usage{
cgsolve(A, b, tol = 1e-06, maxIter = 1000L)
}
\arguments{
\item{A}{matrix, symmetric and positive definite.}

\item{b}{vector, with same dimension as number of rows of A.}

\item{tol}{numeric, threshold for convergence, default is \code{1e-6}.}

\item{maxIter}{numeric, maximum iteration, default is \code{1000}.}
}
\value{
A vector representing solution x.
}
\description{
Conjugate gradient method
}
\details{
Conjugate gradient method for solving system of linear equations Ax = b,
where A is symmetric and positive definite.
}
\examples{
\dontrun{
test_A <- matrix(c(4,1,1,3), ncol = 2)
test_b <- matrix(1:2, ncol = 1)
cgsolve(test_A, test_b, 1e-6, 1000)
}
}
