% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{cgsolve_sparseOMP}
\alias{cgsolve_sparseOMP}
\title{Calculation of sparse matrix inverse and vector product using conjugate gradient descent}
\usage{
cgsolve_sparseOMP(A, b, tol = 1e-06, maxIter = 1000L, nThreads = 1L)
}
\arguments{
\item{A}{A (n x n) matrix.}

\item{b}{A n-vector.}

\item{tol}{Tolerance for conjugate gradient algorithm.}

\item{maxIter}{Max number of iterations for conjugate gradient algorithm.}

\item{nThreads}{Number of threads for OMP.}
}
\value{
A n-vector of A_inv b
}
\description{
This Rcpp function utilizes OpenMP and calculates inverse of a sparse matrix and product with a vector
}
