% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{pcgsolve_sparseOMP}
\alias{pcgsolve_sparseOMP}
\title{Solve for x in Ax = b using preconditioned conjugate gradient method.}
\usage{
pcgsolve_sparseOMP(A, b, preconditioner = "Jacobi", tol = 1e-06,
  maxIter = 1000L, nThreads = 1L)
}
\arguments{
\item{A}{matrix, symmetric and positive definite.}

\item{b}{vector, with same dimension as number of rows of A.}

\item{preconditioner}{string, method for preconditioning: \code{"Jacobi"} (default), \code{"SSOR"}, or \code{"ICC"}.}

\item{tol}{numeric, threshold for convergence, default is \code{1e-6}.}

\item{maxIter}{numeric, maximum iteration, default is \code{1000}.}

\item{nThreads}{Number of threads for OMP.}
}
\value{
A vector representing solution x.
}
\description{
Preconditioned conjugate gradient method
}
\details{
Preconditioned conjugate gradient method for solving system of linear equations Ax = b,
where A is symmetric and positive definite.
}
\examples{
\dontrun{
test_A <- matrix(c(4,1,1,3), ncol = 2)
test_b <- matrix(1:2, ncol = 1)
pcgsolve_sparseOMP(test_A, test_b, "ICC",nThreads=2)
}
}
